tretji zagovor/requirements.txt
flask
numpy
tensorflow
keras
joblib
pandas
pillow
scikit-learn
tensorflow_addons